+++ packages/react-components-migration/react-button-shim-v8-v9/src/index.ts
@@ -1,9 +1,0 @@
-export { shimButtonProps } from './shimButtonProps';
-export { ActionButtonShim } from './ActionButtonShim';
-export { ButtonShim } from './ButtonShim';
-export { ToggleButtonShim } from './ToggleButtonShim';
-export { CommandButtonShim } from './CommandButtonShim';
-export { CompoundButtonShim } from './CompoundButtonShim';
-export { DefaultButtonShim } from './DefaultButtonShim';
-export { PrimaryButtonShim } from './PrimaryButtonShim';
-export { MenuButtonShim } from './MenuButtonShim';