+++ packages/react-components-migration/react-menu-shim-v8-v9/src/stories/MenuShimV8V9/MenuShimDefault.stories.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import { MenuShim, MenuShimProps } from '@fluentui/react-menu-shim-v8-v9';
-
-export const Default = (props: Partial<MenuShimProps>) => <MenuShim {...props} />;