+++ tools/utils.ts
@@ -10,7 +10,7 @@ import {
   ProjectConfiguration,
   readJson,
 } from '@nrwl/devkit';
-import { PackageJson } from './types';
+import { PackageJson, PackageJsonWithBeachball } from './types';
 import * as semver from 'semver';
 
 /**
@@ -186,9 +186,11 @@ export function hasSchemaFlag<T, K extends keyof T>(schema: T, flag: K): schema 
 }
 
 export function isPackageVersionConverged(versionString: string) {
-  const version = semver.parse(versionString);
+  const versionWithoutCaret = versionString.replace('^', '');
+
+  const version = semver.parse(versionWithoutCaret);
   if (version === null) {
-    throw new Error(`${versionString} is not a valid semver version`);
+    throw new Error(`${versionWithoutCaret} is not a valid semver version`);
   }
   return version.major === 9;
 }
@@ -206,4 +208,8 @@ export function isPackageConverged(tree: Tree, project: ProjectConfiguration) {
 export function isV8Package(tree: Tree, project: ProjectConfiguration) {
   const packageJson = readJson<PackageJson>(tree, joinPathFragments(project.root, 'package.json'));
   return packageJson.version.startsWith('8.');
+}
+
+export function packageJsonHasBeachballConfig(packageJson: PackageJson): packageJson is PackageJsonWithBeachball {
+  return !!(packageJson as PackageJsonWithBeachball).beachball;
 }