+++ packages/react-monaco-editor/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-monaco-editor
 
-This log was last generated on Tue, 04 Oct 2022 07:45:03 GMT and should not be manually modified.
+This log was last generated on Mon, 10 Oct 2022 07:38:24 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.7.9](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.7.9)
+
+Mon, 10 Oct 2022 07:38:24 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.7.8..@fluentui/react-monaco-editor_v1.7.9)
+
+### Patches
+
+- Bump @fluentui/react to v8.98.0 ([PR #25138](https://github.com/microsoft/fluentui/pull/25138) by beachball)
 
 ## [1.7.8](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.7.8)